requirements/base.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
openai==0.26.5
pydantic==1.10.5
python-dotenv==1.0.0
requests==2.28.2
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
